+++ packages/vtable-gantt/examples/gantt/gantt-splitLine.ts
@@ -1,13 +1,9 @@
 import type { ColumnsDefine } from '@visactor/vtable';
-import { DateInputEditor, InputEditor } from '@visactor/vtable-editors';
 import type { GanttConstructorOptions, TYPES } from '../../src/index';
-import { Gantt, VTable } from '../../src/index';
+import { Gantt } from '../../src/index';
 import { bindDebugTool } from '../../../vtable/src/scenegraph/debug-tool';
 const CONTAINER_ID = 'vTable';
-const date_input_editor = new DateInputEditor({});
-const input_editor = new InputEditor({});
-VTable.register.editor('input', input_editor);
-VTable.register.editor('date-input', date_input_editor);
+
 export function createTable() {
   const records = [
     {
@@ -779,37 +775,32 @@ export function createTable() {
       field: 'title',
       title: 'title',
       width: 200,
-      sort: true,
-      editor: 'input'
+      sort: true
     },
     {
       field: 'start',
       title: 'start',
       width: 150,
-      sort: true,
-      editor: 'date-input'
+      sort: true
     },
     {
       field: 'end',
       title: 'end',
       width: 150,
-      sort: true,
-      editor: 'date-input'
+      sort: true
     },
     {
       field: 'priority',
       title: 'priority',
       width: 100,
-      sort: true,
-      editor: 'input'
+      sort: true
     },
 
     {
       field: 'progress',
       title: 'progress',
       width: 200,
-      sort: true,
-      editor: 'input'
+      sort: true
     }
   ];
   const option: GanttConstructorOptions = {
@@ -822,10 +813,19 @@ export function createTable() {
     },
 
     frame: {
+      verticalSplitLineMoveable: true,
       outerFrameStyle: {
         borderLineWidth: 2,
         borderColor: 'red',
         cornerRadius: 8
+      },
+      horizontalSplitLine: {
+        lineWidth: 10,
+        lineColor: 'green'
+      },
+      verticalSplitLine: {
+        lineWidth: 20,
+        lineColor: '#e1e4e8'
       },
       verticalSplitLineHighlight: {
         lineColor: 'green',
@@ -845,6 +845,7 @@ export function createTable() {
     },
     headerRowHeight: 60,
     rowHeight: 40,
+
     taskBar: {
       startDateField: 'start',
       endDateField: 'end',
@@ -862,39 +863,21 @@ export function createTable() {
         /** 已完成部分任务条的颜色 */
         completedBarColor: '#91e8e0',
         /** 任务条的圆角 */
-        cornerRadius: 8,
-        /** 任务条的边框 */
-        borderWidth: 1,
-        /** 边框颜色 */
-        borderColor: 'black'
+        cornerRadius: 10
       }
     },
     timelineHeader: {
-      backgroundColor: '#F6F8FA',
       verticalLine: {
         lineWidth: 1,
         lineColor: '#e1e4e8'
       },
       horizontalLine: {
         lineWidth: 1,
-        lineColor: '#e1e4e8'
+        lineColor: 'red'
       },
+      backgroundColor: '#EEF1F5',
+      colWidth: 60,
       scales: [
-        // {
-        //   unit: 'year',
-        //   step: 1,
-        //   format(date: TYPES.DateFormatArgumentType) {
-        //     return `${date.index}`;
-        //   }
-        // },
-
-        // {
-        //   unit: 'month',
-        //   step: 1,
-        //   format(date: TYPES.DateFormatArgumentType) {
-        //     return date.index + '月';
-        //   }
-        // },
         {
           unit: 'week',
           step: 1,
@@ -905,8 +888,7 @@ export function createTable() {
           style: {
             fontSize: 20,
             fontWeight: 'bold',
-            color: 'red',
-            backgroundColor: '#EEF1F5'
+            color: 'red'
           }
         },
         {
@@ -918,13 +900,19 @@ export function createTable() {
           style: {
             fontSize: 20,
             fontWeight: 'bold',
-            color: 'red',
-            backgroundColor: '#EEF1F5'
+            color: 'red'
           }
         }
+        // {
+        //   unit: 'quarter',
+        //   step: 1,
+        //   format(date: TYPES.DateFormatArgumentType) {
+        //     return '第' + date.index + '季度';
+        //   }
+        // }
       ]
     },
-    minDate: '2024-07-07',
+    minDate: '2024-07-01',
     maxDate: '2024-10-15',
     markLine: [
       {
@@ -937,6 +925,8 @@ export function createTable() {
       },
       {
         date: '2024-08-17',
+        position: 'middle',
+        scrollToMarkLine: true,
         style: {
           lineWidth: 2,
           lineColor: 'red',
@@ -955,7 +945,11 @@ export function createTable() {
       style: {
         borderColor: '#e1e4e8'
       }
-    }
+    },
+    scrollStyle: {
+      visible: 'scrolling'
+    },
+    overscrollBehavior: 'none'
   };
   // columns:[
   //   {
@@ -977,11 +971,22 @@ export function createTable() {
   ganttInstance.on('scroll', e => {
     console.log('scroll', e);
   });
-
+  ganttInstance.on('change_date_range', e => {
+    console.log('change_date_range', e);
+  });
+  ganttInstance.on('mouseenter_task_bar', e => {
+    console.log('mouseenter_taskbar', e);
+  });
+  ganttInstance.on('mouseleave_task_bar', e => {
+    console.log('mouseleave_taskbar', e);
+  });
+  ganttInstance.on('click_task_bar', e => {
+    console.log('click_task_bar', e);
+  });
   ganttInstance.listTableInstance?.on('scroll', e => {
     console.log('listTable scroll', e);
   });
-  // bindDebugTool(ganttInstance.scenegraph.stage as any, {
-  //   customGrapicKeys: ['role', '_updateTag']
-  // });
+  bindDebugTool(ganttInstance.scenegraph.stage as any, {
+    customGrapicKeys: ['role', '_updateTag']
+  });
 }