+++ packages/vtable-gantt/src/scenegraph/frame-border.ts
@@ -87,9 +87,13 @@ export class FrameBorder {
       // y: borderLineWidth / 2,
       width:
         this._scene.tableGroup.attribute.width +
-        this.border.attribute.lineWidth +
-        this._scene._gantt.parsedOptions.verticalSplitLine.lineWidth,
-      height: this._scene.tableGroup.attribute.height + this.border.attribute.lineWidth
+        borderLineWidth +
+        (this._scene._gantt.taskListTableInstance ? this._scene._gantt.parsedOptions.verticalSplitLine.lineWidth : 0),
+      // this._scene.tableGroup.attribute.width +
+      // this.border.attribute.lineWidth +
+      // this._scene._gantt.parsedOptions.verticalSplitLine.lineWidth,
+      height: this._scene._gantt.drawHeight + borderLineWidth
+      // height: this._scene.tableGroup.attribute.height + this.border.attribute.lineWidth
     });
   }
 }