+++ packages/vtable/CHANGELOG.md
@@ -1,6 +1,47 @@
 # Change Log - @visactor/vtable
 
-This log was last generated on Mon, 09 Sep 2024 02:02:17 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Sep 2024 08:37:37 GMT and should not be manually modified.
+
+## 1.7.7
+Fri, 13 Sep 2024 08:37:37 GMT
+
+### Updates
+
+- refactor: gantt project export vtable and vrender
+
+
+
+## 1.7.6
+Thu, 12 Sep 2024 06:18:44 GMT
+
+_Version update only_
+
+## 1.7.5
+Wed, 11 Sep 2024 12:33:19 GMT
+
+### Updates
+
+- feat: add getFilteredRecords api #2255
+
+
+- fix: select range click outside not cancel select #2355
+
+
+- fix: split line position fix #2392
+
+
+- fix: levelSpan case front columnNode merge range error #2359
+
+
+- fix: judge value is valid #2402
+
+
+- fix: mousedown chart go dealSelectCell and rerender it #2419
+
+
+- fix: fix axis size and layout #2256
+- fix: fix series number in list-group #2425
+- fix: fix recordIndex config of addRecord in list-group #2426
 
 ## 1.7.4
 Mon, 09 Sep 2024 02:02:17 GMT